pyproject.toml
[project]
# 项目元数据
name = "rm-armor-cnn"
version = "0.1.0"
description = "RoboMaster 装甲板数字识别 CNN 训练框架"
readme = "README.md"

# Python 版本约束：
# 限制在 <3.11
requires-python = ">=3.10, <3.11" 

authors = [
    { name = "Kielas", email = "c1470759@outlook.com" }
]

# 核心运行依赖：项目运行必须安装的包
dependencies = [
    "numpy>=1.24.0,<2.0.0",  # 强制限制在 2.0 以下
    "opencv-python>=4.8.0",
    "pyyaml>=6.0",
    "pathlib>=1.0.1",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "matplotlib>=3.8.0",     
    "ipywidgets>=8.0.0",      
    "ipykernel>=6.15.0",
    "tqdm>=4.66.1",
    "tensorboard>=2.15.0",   # 新增：用于训练过程可视化
]

[project.optional-dependencies]
# 可选依赖：通过 uv sync --extra deploy 安装
# 用于模型导出到 ONNX 并在 OpenVINO/TensorRT 上部署
deploy = [
    "onnx>=1.14.0",
    "onnxscript",      
    "onnxsim>=0.4.33", 
    "torchsummary>=1.5.1",
    "onnxruntime>=1.16.0",   # 新增：用于执行 INT8 动态量化
]

# 开发调试工具：通过 uv sync --extra dev 安装
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
]

[build-system]
# 使用 hatchling 作为构建后端，负责处理包的打包和安装
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# 设置为 ["."] 表示项目源码直接就在根目录下
packages = ["."]

[tool.uv]
# 默认索引：使用清华镜像源加速国内下载
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"

# 额外索引链接：已升级至 CUDA 12.4 专用仓库以兼容 RTX 50 系列显卡 (sm_120)
extra-index-url = ["https://download.pytorch.org/whl/cu124"]

[tool.uv.sources]
# 强制指定 torch 相关包从下方的 pytorch-cu124 索引下载
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }

[[tool.uv.index]]
# 定义名为 pytorch-cu124 的索引
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true